requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
dj-database-url==2.2.0
dj-rest-auth==5.1.0
Django==4.2
django-allauth==0.61.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-filter==23.5
django-location-field==2.7.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
oauthlib==3.2.2
pillow==10.4.0
psycopg2==2.9.9
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2024.2
requests-oauthlib==2.0.0
sqlparse==0.5.1